requirements.txt
streamlit
google-generativeai
google-cloud-discoveryengine
